+++ cg/meta/upload/scout/balsamic_config_builder.py
@@ -55,7 +55,7 @@ class BalsamicConfigBuilder(ScoutConfigBuilder):
         if analysis_type == "tgs":
             analysis_type = "panel"
         if analysis_type == "wgs":
-            raise SyntaxError("Not allowed to upload WGS cancer cases")
+            analysis_type = "wgs"
 
         config_sample.analysis_type = analysis_type
         return config_sample