+++ tests/conftest.py
@@ -1188,10 +1188,9 @@ def fixture_context_config(
         "loqusdb-wes": {"binary_path": "loqusdb", "config_path": "loqusdb-wes-stage.yaml"},
         "balsamic": {
             "root": str(balsamic_dir),
-            "singularity": "BALSAMIC_release_v6.0.1.sif",
-            "reference_config": "reference.json",
             "binary_path": "echo",
             "conda_env": "S_BALSAMIC",
+            "balsamic_cache": "hello",
             "slurm": {
                 "mail_user": "test.email@scilifelab.se",
                 "account": "development",